+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - additional Character methods: CharDirection, GetIntPropertyValue, ToLower, ToTitle, ToUpper
 - partially implemented Normalizer2 class
+- partially implemented ResourceBundle class
 
 ### Changed